Move EndSpanOptions to the main package.
