+++ api/src/main/java/io/opencensus/trace/EndSpanOptions.java
@@ -11,12 +11,12 @@
  * limitations under the License.
  */
 
-package io.opencensus.trace.base;
+package io.opencensus.trace;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
 import com.google.auto.value.AutoValue;
-import io.opencensus.trace.Span;
+import io.opencensus.trace.base.Status;
 import javax.annotation.concurrent.Immutable;
 
 /**
