+++ api/src/main/java/io/opencensus/trace/Span.java
@@ -18,7 +18,6 @@ import static com.google.common.base.Preconditions.checkNotNull;
 
 import io.opencensus.trace.base.Annotation;
 import io.opencensus.trace.base.AttributeValue;
-import io.opencensus.trace.base.EndSpanOptions;
 import io.opencensus.trace.base.Link;
 import io.opencensus.trace.base.NetworkEvent;
 import java.util.Collections;
