+++ api/src/main/java/io/opencensus/trace/SpanBuilder.java
@@ -16,7 +16,6 @@ package io.opencensus.trace;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 import io.opencensus.common.NonThrowingCloseable;
-import io.opencensus.trace.base.EndSpanOptions;
 import java.util.List;
 import javax.annotation.Nullable;
