+++ api/src/test/java/io/opencensus/trace/BlankSpanTest.java
@@ -17,7 +17,6 @@ import static com.google.common.truth.Truth.assertThat;
 
 import io.opencensus.trace.base.Annotation;
 import io.opencensus.trace.base.AttributeValue;
-import io.opencensus.trace.base.EndSpanOptions;
 import io.opencensus.trace.base.Link;
 import io.opencensus.trace.base.NetworkEvent;
 import java.util.HashMap;
