+++ api/src/test/java/io/opencensus/trace/ScopedSpanHandleTest.java
@@ -18,7 +18,6 @@ import static org.mockito.Matchers.same;
 import static org.mockito.Mockito.verify;
 
 import io.opencensus.common.NonThrowingCloseable;
-import io.opencensus.trace.base.EndSpanOptions;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
