+++ impl_core/src/main/java/io/opencensus/trace/SpanImpl.java
@@ -22,7 +22,6 @@ import io.opencensus.common.Clock;
 import io.opencensus.internal.TimestampConverter;
 import io.opencensus.trace.base.Annotation;
 import io.opencensus.trace.base.AttributeValue;
-import io.opencensus.trace.base.EndSpanOptions;
 import io.opencensus.trace.base.Link;
 import io.opencensus.trace.base.NetworkEvent;
 import io.opencensus.trace.base.SpanId;
