+++ impl_core/src/test/java/io/opencensus/trace/export/SampledSpanStoreImplTest.java
@@ -18,12 +18,12 @@ import static com.google.common.truth.Truth.assertThat;
 import io.opencensus.common.Duration;
 import io.opencensus.common.Timestamp;
 import io.opencensus.testing.common.TestClock;
+import io.opencensus.trace.EndSpanOptions;
 import io.opencensus.trace.Span;
 import io.opencensus.trace.Span.Options;
 import io.opencensus.trace.SpanContext;
 import io.opencensus.trace.SpanImpl;
 import io.opencensus.trace.SpanImpl.StartEndHandler;
-import io.opencensus.trace.base.EndSpanOptions;
 import io.opencensus.trace.base.SpanId;
 import io.opencensus.trace.base.Status;
 import io.opencensus.trace.base.Status.CanonicalCode;
